Organizing so app.py stays in a readable state
